Move user confirmation from resolver to controller
